Remove server bundle since the agnostic core doesn't depend on React
