--- packages/next-intl/server.d.ts
@@ -1,1 +1,0 @@
-export * from './dist/server';
