--- packages/use-intl/server.js
@@ -1,1 +1,0 @@
-module.exports = require('./dist/server');
