--- packages/use-intl/src/server/index.tsx
@@ -1,3 +1,0 @@
-// Public API of the `server` package
-export {default as createIntl} from '../core/createIntl';
-export {default as createTranslator} from './createTranslator';
